requirements.txt
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
Deprecated==1.2.4
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.15.4
PyGithub==1.43.4
PyJWT==1.7.1
pyparsing==2.3.0
python-dateutil==2.7.5
requests==2.21.0
six==1.12.0
smmap2==2.0.5
urllib3==1.24.1
wrapt==1.10.11
